+++ sentry-spring-boot-starter/src/main/java/io/sentry/spring/boot/SentryWebfluxAutoConfiguration.java
@@ -5,16 +5,13 @@ import io.sentry.IHub;
 import io.sentry.spring.webflux.SentryScheduleHook;
 import io.sentry.spring.webflux.SentryWebExceptionHandler;
 import io.sentry.spring.webflux.SentryWebFilter;
-import io.sentry.spring.webflux.SentryWebTracingFilter;
 import org.jetbrains.annotations.ApiStatus;
 import org.jetbrains.annotations.NotNull;
 import org.springframework.boot.ApplicationRunner;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnBean;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnClass;
-import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnWebApplication;
 import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Conditional;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.core.Ordered;
 import org.springframework.core.annotation.Order;
@@ -43,14 +40,6 @@ public class SentryWebfluxAutoConfiguration {
   @Order(SENTRY_SPRING_FILTER_PRECEDENCE)
   public @NotNull SentryWebFilter sentryWebFilter(final @NotNull IHub hub) {
     return new SentryWebFilter(hub);
-  }
-
-  @Bean
-  @Order(SENTRY_SPRING_FILTER_PRECEDENCE + 1)
-  @Conditional(SentryAutoConfiguration.SentryTracingCondition.class)
-  @ConditionalOnMissingBean(name = "sentryWebTracingFilter")
-  public @NotNull SentryWebTracingFilter sentryWebTracingFilter() {
-    return new SentryWebTracingFilter();
   }
 
   /** Configures exception handler that handles unhandled exceptions and sends them to Sentry. */