+++ sentry-spring/src/test/kotlin/io/sentry/spring/webflux/SentryWebFluxTracingFilterTest.kt
@@ -1,6 +1,10 @@
 package io.sentry.spring.webflux
 
+import io.sentry.Breadcrumb
+import io.sentry.Hint
 import io.sentry.IHub
+import io.sentry.ScopeCallback
+import io.sentry.Sentry
 import io.sentry.SentryOptions
 import io.sentry.SentryTracer
 import io.sentry.SpanId
@@ -12,10 +16,12 @@ import io.sentry.protocol.SentryId
 import io.sentry.protocol.TransactionNameSource
 import io.sentry.spring.webflux.SentryWebFilter.SENTRY_HUB_KEY
 import org.assertj.core.api.Assertions.assertThat
+import org.mockito.Mockito
 import org.mockito.kotlin.any
 import org.mockito.kotlin.anyOrNull
 import org.mockito.kotlin.check
 import org.mockito.kotlin.mock
+import org.mockito.kotlin.times
 import org.mockito.kotlin.verify
 import org.mockito.kotlin.verifyNoMoreInteractions
 import org.mockito.kotlin.whenever
@@ -49,7 +55,7 @@ class SentryWebFluxTracingFilterTest {
             whenever(hub.options).thenReturn(options)
         }
 
-        fun getSut(isEnabled: Boolean = true, status: HttpStatus = HttpStatus.OK, sentryTraceHeader: String? = null, method: HttpMethod = HttpMethod.POST): SentryWebTracingFilter {
+        fun getSut(isEnabled: Boolean = true, status: HttpStatus = HttpStatus.OK, sentryTraceHeader: String? = null, method: HttpMethod = HttpMethod.POST): SentryWebFilter {
             var requestBuilder = MockServerHttpRequest.method(method, "/product/{id}", 12)
             if (sentryTraceHeader != null) {
                 requestBuilder = requestBuilder.header("sentry-trace", sentryTraceHeader)
@@ -63,89 +69,102 @@ class SentryWebFluxTracingFilterTest {
             whenever(hub.startTransaction(any(), check<TransactionOptions> { assertTrue(it.isBindToScope) })).thenAnswer { SentryTracer(it.arguments[0] as TransactionContext, hub) }
             whenever(hub.isEnabled).thenReturn(isEnabled)
             whenever(chain.filter(any())).thenReturn(Mono.create { s -> s.success() })
-            return SentryWebTracingFilter()
+            return SentryWebFilter(hub)
         }
     }
 
     private val fixture = Fixture()
 
+    fun withMockHub(closure: () -> Unit) = Mockito.mockStatic(Sentry::class.java).use {
+        it.`when`<Any> { Sentry.cloneMainHub() }.thenReturn(fixture.hub)
+        closure.invoke()
+    }
+
     @Test
     fun `creates transaction around the request`() {
         val filter = fixture.getSut()
 
-        filter.filter(fixture.exchange, fixture.chain).block()
+        withMockHub {
+            filter.filter(fixture.exchange, fixture.chain).block()
 
-        verify(fixture.hub).startTransaction(
-            check<TransactionContext> {
-                assertEquals("POST /product/12", it.name)
-                assertEquals(TransactionNameSource.URL, it.transactionNameSource)
-                assertEquals("http.server", it.operation)
-            },
-            check<TransactionOptions> {
-                assertNotNull(it.customSamplingContext?.get("request"))
-                assertTrue(it.customSamplingContext?.get("request") is ServerHttpRequest)
-                assertTrue(it.isBindToScope)
-            }
-        )
-        verify(fixture.chain).filter(fixture.exchange)
-        verify(fixture.hub).captureTransaction(
-            check {
-                assertThat(it.transaction).isEqualTo("POST /product/{id}")
-                assertThat(it.contexts.trace!!.status).isEqualTo(SpanStatus.OK)
-                assertThat(it.contexts.trace!!.operation).isEqualTo("http.server")
-            },
-            anyOrNull<TraceContext>(),
-            anyOrNull(),
-            anyOrNull()
-        )
+            verify(fixture.hub).startTransaction(
+                check<TransactionContext> {
+                    assertEquals("POST /product/12", it.name)
+                    assertEquals(TransactionNameSource.URL, it.transactionNameSource)
+                    assertEquals("http.server", it.operation)
+                },
+                check<TransactionOptions> {
+                    assertNotNull(it.customSamplingContext?.get("request"))
+                    assertTrue(it.customSamplingContext?.get("request") is ServerHttpRequest)
+                    assertTrue(it.isBindToScope)
+                }
+            )
+            verify(fixture.chain).filter(fixture.exchange)
+            verify(fixture.hub).captureTransaction(
+                check {
+                    assertThat(it.transaction).isEqualTo("POST /product/{id}")
+                    assertThat(it.contexts.trace!!.status).isEqualTo(SpanStatus.OK)
+                    assertThat(it.contexts.trace!!.operation).isEqualTo("http.server")
+                },
+                anyOrNull<TraceContext>(),
+                anyOrNull(),
+                anyOrNull()
+            )
+        }
     }
 
     @Test
     fun `sets correct span status based on the response status`() {
         val filter = fixture.getSut(status = HttpStatus.INTERNAL_SERVER_ERROR)
 
-        filter.filter(fixture.exchange, fixture.chain).block()
+        withMockHub {
+            filter.filter(fixture.exchange, fixture.chain).block()
 
-        verify(fixture.hub).captureTransaction(
-            check {
-                assertThat(it.contexts.trace!!.status).isEqualTo(SpanStatus.INTERNAL_ERROR)
-            },
-            anyOrNull<TraceContext>(),
-            anyOrNull(),
-            anyOrNull()
-        )
+            verify(fixture.hub).captureTransaction(
+                check {
+                    assertThat(it.contexts.trace!!.status).isEqualTo(SpanStatus.INTERNAL_ERROR)
+                },
+                anyOrNull<TraceContext>(),
+                anyOrNull(),
+                anyOrNull()
+            )
+        }
     }
 
     @Test
     fun `does not set span status for response status that dont match predefined span statuses`() {
         val filter = fixture.getSut(status = HttpStatus.FOUND)
 
-        filter.filter(fixture.exchange, fixture.chain).block()
+        withMockHub {
+            filter.filter(fixture.exchange, fixture.chain).block()
 
-        verify(fixture.hub).captureTransaction(
-            check {
-                assertThat(it.contexts.trace!!.status).isNull()
-            },
-            anyOrNull<TraceContext>(),
-            anyOrNull(),
-            anyOrNull()
-        )
+            verify(fixture.hub).captureTransaction(
+                check {
+                    assertThat(it.contexts.trace!!.status).isNull()
+                },
+                anyOrNull<TraceContext>(),
+                anyOrNull(),
+                anyOrNull()
+            )
+        }
     }
 
     @Test
     fun `when sentry trace is not present, transaction does not have parentSpanId set`() {
         val filter = fixture.getSut()
 
-        filter.filter(fixture.exchange, fixture.chain).block()
+        withMockHub {
+            filter.filter(fixture.exchange, fixture.chain).block()
 
-        verify(fixture.hub).captureTransaction(
-            check {
-                assertThat(it.contexts.trace!!.parentSpanId).isNull()
-            },
-            anyOrNull<TraceContext>(),
-            anyOrNull(),
-            anyOrNull()
-        )
+            verify(fixture.hub).captureTransaction(
+                check {
+                    assertThat(it.contexts.trace!!.parentSpanId).isNull()
+                },
+                anyOrNull<TraceContext>(),
+                anyOrNull(),
+                anyOrNull()
+            )
+        }
     }
 
     @Test
@@ -153,99 +172,119 @@ class SentryWebFluxTracingFilterTest {
         val parentSpanId = SpanId()
         val filter = fixture.getSut(sentryTraceHeader = "${SentryId()}-$parentSpanId-1")
 
-        filter.filter(fixture.exchange, fixture.chain).block()
+        withMockHub {
+            filter.filter(fixture.exchange, fixture.chain).block()
 
-        verify(fixture.hub).captureTransaction(
-            check {
-                assertThat(it.contexts.trace!!.parentSpanId).isEqualTo(parentSpanId)
-            },
-            anyOrNull<TraceContext>(),
-            anyOrNull(),
-            anyOrNull()
-        )
+            verify(fixture.hub).captureTransaction(
+                check {
+                    assertThat(it.contexts.trace!!.parentSpanId).isEqualTo(parentSpanId)
+                },
+                anyOrNull<TraceContext>(),
+                anyOrNull(),
+                anyOrNull()
+            )
+        }
     }
 
     @Test
     fun `when hub is disabled, components are not invoked`() {
         val filter = fixture.getSut(isEnabled = false)
 
-        filter.filter(fixture.exchange, fixture.chain).block()
+        withMockHub {
+            filter.filter(fixture.exchange, fixture.chain).block()
 
-        verify(fixture.chain).filter(fixture.exchange)
+            verify(fixture.chain).filter(fixture.exchange)
 
-        verify(fixture.hub).isEnabled
-        verifyNoMoreInteractions(fixture.hub)
+            verify(fixture.hub).isEnabled
+            verifyNoMoreInteractions(fixture.hub)
+        }
     }
 
     @Test
     fun `sets status to internal server error when chain throws exception`() {
         val filter = fixture.getSut()
-        whenever(fixture.chain.filter(any())).thenReturn(Mono.error(RuntimeException("error")))
 
-        try {
-            filter.filter(fixture.exchange, fixture.chain).block()
-            fail("filter is expected to rethrow exception")
-        } catch (_: Exception) {
+        withMockHub {
+            whenever(fixture.chain.filter(any())).thenReturn(Mono.error(RuntimeException("error")))
+
+            try {
+                filter.filter(fixture.exchange, fixture.chain).block()
+                fail("filter is expected to rethrow exception")
+            } catch (_: Exception) {
+            }
+            verify(fixture.hub).captureTransaction(
+                check {
+                    assertThat(it.status).isEqualTo(SpanStatus.INTERNAL_ERROR)
+                },
+                anyOrNull<TraceContext>(),
+                anyOrNull(),
+                anyOrNull()
+            )
         }
-        verify(fixture.hub).captureTransaction(
-            check {
-                assertThat(it.status).isEqualTo(SpanStatus.INTERNAL_ERROR)
-            },
-            anyOrNull<TraceContext>(),
-            anyOrNull(),
-            anyOrNull()
-        )
     }
 
     @Test
     fun `does not track OPTIONS request with traceOptionsRequests=false`() {
         val filter = fixture.getSut(method = HttpMethod.OPTIONS)
-        fixture.options.isTraceOptionsRequests = false
 
-        filter.filter(fixture.exchange, fixture.chain).block()
+        withMockHub {
+            fixture.options.isTraceOptionsRequests = false
 
-        verify(fixture.chain).filter(fixture.exchange)
+            filter.filter(fixture.exchange, fixture.chain).block()
 
-        verify(fixture.hub).isEnabled
-        verify(fixture.hub).options
-        verifyNoMoreInteractions(fixture.hub)
+            verify(fixture.chain).filter(fixture.exchange)
+
+            verify(fixture.hub).isEnabled
+            verify(fixture.hub, times(2)).options
+            verify(fixture.hub).pushScope()
+            verify(fixture.hub).addBreadcrumb(any<Breadcrumb>(), any<Hint>())
+            verify(fixture.hub).configureScope(any<ScopeCallback>())
+            verify(fixture.hub).popScope()
+            verifyNoMoreInteractions(fixture.hub)
+        }
     }
 
     @Test
     fun `tracks OPTIONS request with traceOptionsRequests=true`() {
         val filter = fixture.getSut(method = HttpMethod.OPTIONS)
-        fixture.options.isTraceOptionsRequests = true
 
-        filter.filter(fixture.exchange, fixture.chain).block()
+        withMockHub {
+            fixture.options.isTraceOptionsRequests = true
 
-        verify(fixture.chain).filter(fixture.exchange)
+            filter.filter(fixture.exchange, fixture.chain).block()
 
-        verify(fixture.hub).captureTransaction(
-            check {
-                assertThat(it.contexts.trace!!.parentSpanId).isNull()
-            },
-            anyOrNull<TraceContext>(),
-            anyOrNull(),
-            anyOrNull()
-        )
+            verify(fixture.chain).filter(fixture.exchange)
+
+            verify(fixture.hub).captureTransaction(
+                check {
+                    assertThat(it.contexts.trace!!.parentSpanId).isNull()
+                },
+                anyOrNull<TraceContext>(),
+                anyOrNull(),
+                anyOrNull()
+            )
+        }
     }
 
     @Test
     fun `tracks POST request with traceOptionsRequests=false`() {
         val filter = fixture.getSut(method = HttpMethod.POST)
-        fixture.options.isTraceOptionsRequests = false
 
-        filter.filter(fixture.exchange, fixture.chain).block()
+        withMockHub {
+            fixture.options.isTraceOptionsRequests = false
 
-        verify(fixture.chain).filter(fixture.exchange)
+            filter.filter(fixture.exchange, fixture.chain).block()
 
-        verify(fixture.hub).captureTransaction(
-            check {
-                assertThat(it.contexts.trace!!.parentSpanId).isNull()
-            },
-            anyOrNull<TraceContext>(),
-            anyOrNull(),
-            anyOrNull()
-        )
+            verify(fixture.chain).filter(fixture.exchange)
+
+            verify(fixture.hub).captureTransaction(
+                check {
+                    assertThat(it.contexts.trace!!.parentSpanId).isNull()
+                },
+                anyOrNull<TraceContext>(),
+                anyOrNull(),
+                anyOrNull()
+            )
+        }
     }
 }